feat(web): the Ask verb folds into the one box
ask.html is gone. The two-request stream, its five routes and the seven
template structs move to workspace.rs verbatim; only the door is new.
ui.rs drops below ten thousand lines for the first time.

The excerpts land in #results, the same rail the search results were
in. The rail holds what the current act produced, and an ask is a
different act from the search before it.

The last-query prefill goes. It existed because a query typed on the
rail and then retyped into ask was the cost of two pages with nothing
carried between them; there is one box now and the query is already in
it. a_query_typed_on_search_arrives_in_the_ask_box went with it.

Three tests followed the code they cover into workspace.rs, including
the one that reads sse_event's own source to check the driver handles
every frame the server names.

tests/browser/ask_stream.js is updated to the new markup but NOT run:
it needs node and a headless Chrome, and it fails on this machine — on
79612e1 as well as here, so this change is not what broke it. The Rust
suite (1,623) and clippy are green.
